+++ src/main/java/org/zhadok/poe/controller/ControllerEventListener.java
@@ -6,4 +6,6 @@ public interface ControllerEventListener {
 
 	void handleEvent(Event event);
 
+	void handleApplicationExit(); 
+	
 }