+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 - 0.0.16
     - Add cancel button while listening for next controller input/output
-    - 
+    - App window remembers resizing between restarts
 - 0.0.14
     - Add poe-controller version to application title
     - Change default start method via .jar (not .cmd)